scenenet20/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
ansi2html=1.9.1=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
brotli-python=1.0.9=py310h6a678d5_7
build=1.0.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.3.11=h06a4308_0
cachecontrol=0.13.1=pypi_0
cachetools=5.3.2=pypi_0
ccimport=0.4.2=pypi_0
certifi=2022.12.7=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=2.1.1=pypi_0
cleo=2.1.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.1=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.0=pypi_0
crashtest=0.4.1=pypi_0
cryptography=41.0.7=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-runtime=12.1.0=0
cumm-cu120=0.4.11=pypi_0
cycler=0.12.1=pypi_0
dash=2.14.2=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
decorator=5.1.1=pypi_0
distlib=0.3.8=pypi_0
docker-pycreds=0.4.0=pypi_0
dulwich=0.21.7=pypi_0
easydict=1.11=pypi_0
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fire=0.5.0=pypi_0
flash-attn=2.5.5=pypi_0
flask=3.0.1=pypi_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.24.0=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
grpcio=1.60.0rc1=pypi_0
h5py=3.11.0=pypi_0
huggingface-hub=0.20.3=pypi_0
idna=3.4=py310h06a4308_0
imbalanced-learn=0.12.2=pypi_0
imblearn=0.0=pypi_0
importlib-metadata=7.0.1=pypi_0
installer=0.7.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
ipdb=0.13.13=pypi_0
ipython=8.20.0=pypi_0
ipywidgets=8.1.1=pypi_0
itsdangerous=2.1.2=pypi_0
jaraco-classes=3.3.0=pypi_0
jedi=0.19.1=pypi_0
jeepney=0.8.0=pypi_0
jinja2=3.1.3=py310h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-core=5.7.1=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
keopscore=2.1.2=pypi_0
keyring=24.3.0=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lark=1.1.9=pypi_0
laszip=0.2.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.8.1.2=0
libcurand=10.3.4.107=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lightning=2.1.2=pypi_0
lightning-utilities=0.11.2=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
loguru=0.7.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5.1=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
more-itertools=10.2.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
msgpack=1.0.7=pypi_0
multidict=6.0.5=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.1.105=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
ocnn=2.2.1=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_2
packaging=24.0=pypi_0
pandas=2.2.0=pypi_0
parso=0.8.3=pypi_0
pccm=0.4.11=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=py310h5eee18b_0
pip=23.3.1=py310h06a4308_0
pkginfo=1.9.6=pypi_0
platformdirs=4.1.0=pypi_0
plotly=5.18.0=pypi_0
poetry=1.7.1=pypi_0
poetry-core=1.8.1=pypi_0
poetry-plugin-export=1.6.0=pypi_0
pointops=1.0=dev_0
pointops2=1.0=dev_0
portalocker=2.8.2=pypi_0
prettytable=3.9.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pybind11=2.11.1=pypi_0
pygments=2.17.2=pypi_0
pykeops=2.1.2=pypi_0
pyparsing=3.1.2=pypi_0
pyproject-hooks=1.0.0=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=2.1.0=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cluster=1.6.3=py310_torch_2.1.0_cu121
pytorch-cuda=12.1=ha16c6d3_5
pytorch-lightning=2.2.5=pypi_0
pytorch-mutex=1.0=cuda
pytorch-scatter=2.1.2=py310_torch_2.1.0_cu121
pytorch-sparse=0.6.18=py310_torch_2.1.0_cu121
pytorch3d=0.7.6=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=py310h5eee18b_0
rapidfuzz=3.6.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.32.1=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=1.0.0=pypi_0
retrying=1.3.4=pypi_0
rpds-py=0.17.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
secretstorage=3.3.3=pypi_0
sentry-sdk=1.37.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=70.0.0=pypi_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
spconv-cu120=2.3.6=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=py310h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h1ccaba5_0
tomlkit=0.12.3=pypi_0
torch=2.2.0+cu121=pypi_0
torch-cluster=1.6.3+pt21cu121=pypi_0
torch-geometric=2.5.3=pypi_0
torchaudio=2.2.0+cu121=pypi_0
torchinfo=1.8.0=pypi_0
torchmetrics=1.4.0.post0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.17.0+cu121=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.1=pypi_0
triton=2.3.0=pypi_0
trove-classifiers=2024.1.8=pypi_0
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2023.4=pypi_0
urllib3=1.26.13=pypi_0
virtualenv=20.25.0=pypi_0
wandb=0.16.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
